backend/requirements.txt
Django==5.2.7
djangorestframework==3.15.2
django-cors-headers==4.6.0